refactor: [#1478] move EventBus to event mod
