+++ packages/http-tracker-core/benches/helpers/util.rs
@@ -3,10 +3,10 @@ use std::sync::Arc;
 
 use aquatic_udp_protocol::{AnnounceEvent, NumberOfBytes, PeerId};
 use bittorrent_http_tracker_core::event;
+use bittorrent_http_tracker_core::event::bus::EventBus;
 use bittorrent_http_tracker_core::event::sender::Broadcaster;
 use bittorrent_http_tracker_core::event::Event;
 use bittorrent_http_tracker_core::statistics::event::listener::run_event_listener;
-use bittorrent_http_tracker_core::statistics::event_bus::EventBus;
 use bittorrent_http_tracker_core::statistics::repository::Repository;
 use bittorrent_http_tracker_protocol::v1::requests::announce::Announce;
 use bittorrent_http_tracker_protocol::v1::services::peer_ip_resolver::ClientIpSources;
