+++ packages/udp-tracker-core/src/event/mod.rs
@@ -1,3 +1,6 @@
+pub mod bus;
+pub mod sender;
+
 use std::net::SocketAddr;
 
 use bittorrent_primitives::info_hash::InfoHash;
@@ -5,8 +8,6 @@ use torrust_tracker_metrics::label::{LabelSet, LabelValue};
 use torrust_tracker_metrics::label_name;
 use torrust_tracker_primitives::peer::PeerAnnouncement;
 use torrust_tracker_primitives::service_binding::ServiceBinding;
-
-pub mod sender;
 
 /// A UDP core event.
 #[derive(Debug, PartialEq, Eq, Clone)]
