+++ packages/udp-tracker-core/src/statistics/mod.rs
@@ -1,5 +1,4 @@
 pub mod event;
-pub mod event_bus;
 pub mod metrics;
 pub mod repository;
 pub mod services;
