+++ packages/udp-tracker-server/src/handlers/announce.rs
@@ -375,10 +375,7 @@ mod tests {
                 core_udp_tracker_services: Arc<CoreUdpTrackerServices>,
             ) -> Response {
                 let udp_server_broadcaster = crate::event::sender::Broadcaster::default();
-                let event_bus = Arc::new(crate::statistics::event_bus::EventBus::new(
-                    false,
-                    udp_server_broadcaster.clone(),
-                ));
+                let event_bus = Arc::new(crate::event::bus::EventBus::new(false, udp_server_broadcaster.clone()));
 
                 let udp_server_stats_event_sender = event_bus.sender();
 
@@ -536,9 +533,9 @@ mod tests {
             use bittorrent_tracker_core::torrent::repository::in_memory::InMemoryTorrentRepository;
             use bittorrent_tracker_core::whitelist;
             use bittorrent_udp_tracker_core::connection_cookie::{gen_remote_fingerprint, make};
+            use bittorrent_udp_tracker_core::event::bus::EventBus;
             use bittorrent_udp_tracker_core::event::sender::Broadcaster;
             use bittorrent_udp_tracker_core::services::announce::AnnounceService;
-            use bittorrent_udp_tracker_core::statistics::event_bus::EventBus;
             use mockall::predicate::eq;
             use torrust_tracker_configuration::Core;
             use torrust_tracker_primitives::service_binding::{Protocol, ServiceBinding};
@@ -718,10 +715,7 @@ mod tests {
                 let udp_core_stats_event_sender = core_event_bus.sender();
 
                 let udp_server_broadcaster = crate::event::sender::Broadcaster::default();
-                let server_event_bus = Arc::new(crate::statistics::event_bus::EventBus::new(
-                    false,
-                    udp_server_broadcaster.clone(),
-                ));
+                let server_event_bus = Arc::new(crate::event::bus::EventBus::new(false, udp_server_broadcaster.clone()));
 
                 let udp_server_stats_event_sender = server_event_bus.sender();
